refactor: create domain/view package and divide codes
