+++ src/main/java/racingcar/domain/Game.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.List;
 import java.util.Random;
