+++ src/main/java/racingcar/view/InputHandler.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.ArrayList;
 import java.util.Arrays;
