+++ src/main/java/racingcar/view/OutputHandler.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Car;
 
 import java.util.List;
