+++ src/test/java/racingcar/GameTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarGenerator;
+import racingcar.domain.Game;
 
 import static org.assertj.core.api.Assertions.*;
