+++ src/test/java/racingcar/InputHandlerTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.view.InputHandler;
 
 import static org.assertj.core.api.Assertions.*;
